backend/requirements.txt
flask
flask-cors
pandas
tensorflow-cpu
scikit-learn
